refactor(streaming): extract the warm executor into lookup/streaming_warm.py (LML#1121)
The review's F1/F3 pass left lookup/streaming_url_postprocess.py at 870 lines against its 800 ceiling, and 800 was arithmetically unreachable: the file measured 781 on main, and the PR adds 23 executable lines, so the floor sat above the ceiling before a word of prose. This table's convention is to extract rather than raise, with direct precedent in streaming_warm_admission.py -- carved out of this same file for this same reason.

The seam is the one already implicit in the module: streaming_url_postprocess.py decides whether to warm, streaming_warm.py executes one. The concurrency semaphore, the in-flight dedup set, the background-task registry, the enqueue, the warm coroutine, its probe-timeout resolver and its release-identity mint move together, since every one of them is state or logic the executor owns. StreamingUrlCacheConfig is imported under TYPE_CHECKING to keep the dependency one-way.

Pure motion: every moved function body is byte-identical to its original apart from the prose dedupe already applied in this PR. Both import orders were verified to succeed, since a cycle that only bites depending on which module loads first survives inspection and dies at startup.
